feat: enhance COC modules with video integration and mobile responsiveness
- Add YouTube video embeds to all COC modules (COC1-4) with uniform purple/blue button styling
- Implement body scroll locking when modals are open to prevent background scrolling
- Fix mobile responsiveness issues across all module components and modals
- Update COC1 modules 3-4 with new content: 'Creating Bootable Media' and 'Operating System Installation'
- Add comprehensive useful links section to glossary page with clickable resources
- Standardize button colors: purple for video, blue for practice
