+++ src/app/coc2/page.tsx
@@ -1,12 +1,9 @@
-"use client"
-
 import Link from 'next/link'
+import { Metadata } from 'next'
 import { 
   Network, 
   ArrowLeft, 
-  BookOpen, 
-  Video, 
-  FileText, 
+  BookOpen,
   CheckCircle,
   Clock,
   Users,
@@ -16,6 +13,35 @@ import {
   Server
 } from 'lucide-react'
 import Footer from '@/components/Footer'
+import COC2Modules from './COC2Modules'
+
+export const metadata: Metadata = {
+  title: 'COC 2: Setting up Computer Networks - CSS NC II Training',
+  description: 'Master wired and wireless network setup, cable creation, router configuration, and workstation connectivity. Learn network protocols, IP addressing, and network troubleshooting.',
+  keywords: 'COC 2, Computer Networks, Network Setup, Cable Creation, Router Configuration, Workstation Connectivity, Network Protocols, IP Addressing, CSS NC II, TESDA',
+  openGraph: {
+    title: 'COC 2: Setting up Computer Networks - CSS NC II Training',
+    description: 'Master wired and wireless network setup, cable creation, router configuration, and workstation connectivity.',
+    url: 'https://css-nc2-companion.vercel.app/coc2',
+    siteName: 'CSS NC II Training Companion App',
+    images: [
+      {
+        url: '/img/meta-coc2.jpg',
+        width: 1200,
+        height: 630,
+        alt: 'COC 2: Setting up Computer Networks',
+      },
+    ],
+    locale: 'en_US',
+    type: 'website',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'COC 2: Setting up Computer Networks - CSS NC II Training',
+    description: 'Master wired and wireless network setup, cable creation, router configuration, and workstation connectivity.',
+    images: ['/img/meta-coc2.jpg'],
+  },
+};
 
 export default function COC2Page() {
   const learningObjectives = [
@@ -25,206 +51,6 @@ export default function COC2Page() {
     'Set up Wi-Fi and wireless repeaters',
     'Inspect and test configured networks',
     'Troubleshoot network connectivity issues'
-  ]
-
-  const modules = [
-    {
-      id: 'module1',
-      title: 'Network Fundamentals',
-      description: 'Understanding computer networks, types, and topology',
-      duration: '6 hours',
-      videoUrl: 'https://www.youtube.com/watch?v=EWTJKcg7Pj8',
-      topics: ['Network Types (LAN, WAN, MAN)', 'Network Topology', 'Network Components', 'OSI Model'],
-      practiceContent: {
-        title: 'Network Types and Topology',
-        type: 'A',
-        objectives: 'Understand different network types and topology configurations',
-        materials: ['Network diagrams', 'Computer systems', 'Network simulation software'],
-        procedures: [
-          'Study different network types (LAN, WAN, PAN, MAN)',
-          'Analyze network topology diagrams',
-          'Identify network components and their functions',
-          'Practice network design principles',
-          'Document network architecture'
-        ],
-        networkTypes: [
-          {
-            name: 'Local Area Networks (LAN)',
-            description: 'Network that connects computers within a limited area',
-            examples: ['School network', 'Office network', 'Home network']
-          },
-          {
-            name: 'Wide Area Networks (WAN)',
-            description: 'Network that covers larger geographical areas',
-            examples: ['Internet', 'Corporate network across cities']
-          },
-          {
-            name: 'Personal Area Networks (PAN)',
-            description: 'Network based on individual workspace',
-            examples: ['Bluetooth devices', 'Personal computer setup']
-          },
-          {
-            name: 'Metropolitan Area Networks (MAN)',
-            description: 'Network stretching across metropolitan areas',
-            examples: ['City-wide network', 'University campus network']
-          }
-        ],
-        topologyTypes: [
-          {
-            name: 'Star Topology',
-            description: 'Central node connects to each computer independently',
-            advantages: ['Easy to manage', 'One failure doesn&apos;t affect others'],
-            disadvantages: ['Requires many cables', 'Central point failure']
-          },
-          {
-            name: 'Bus Topology',
-            description: 'Single cable connects all computers',
-            advantages: ['Less cabling required', 'Simple setup'],
-            disadvantages: ['Single point of failure', 'Difficult troubleshooting']
-          },
-          {
-            name: 'Ring Topology',
-            description: 'Computers connected in a circular pattern',
-            advantages: ['Equal access to network', 'Good performance'],
-            disadvantages: ['Single failure affects all', 'Difficult to expand']
-          },
-          {
-            name: 'Mesh Topology',
-            description: 'All nodes interconnected with each other',
-            advantages: ['Highly reliable', 'Multiple paths available'],
-            disadvantages: ['Expensive', 'Complex configuration']
-          }
-        ]
-      }
-    },
-    {
-      id: 'module2',
-      title: 'Network Cabling',
-      description: 'Cable types, crimping, and network tools',
-      duration: '8 hours',
-      videoUrl: 'https://www.youtube.com/watch?v=NWhoJp8UQpo',
-      topics: ['UTP/STP Cables', 'Cable Crimping', 'Color Coding', 'Network Tools'],
-      practiceContent: {
-        title: 'Cable Crimping with Color Coding',
-        type: 'A',
-        objectives: 'Create cable connection using UTP cable with specific standard tools and materials in accordance to cable color coding',
-        materials: ['UTP Cable', 'RJ45 Connectors', 'Crimping Tools', 'LAN Cable Tester'],
-        procedures: [
-          'Get a UTP cable and crimp both ends using crimping tools and RJ45',
-          'Perform color coding T-568A',
-          'Test the cable using LAN cable tester'
-        ],
-        cableTypes: [
-          {
-            name: 'Unshielded Twisted Pair (UTP)',
-            description: 'Copper cabling with twisted pairs without shielding',
-            categories: ['Cat5e', 'Cat6', 'Cat6a', 'Cat7'],
-            uses: ['Ethernet networks', 'Telephone systems', 'Data transmission']
-          },
-          {
-            name: 'Shielded Twisted Pair (STP)',
-            description: 'Copper cabling with additional shielding',
-            advantages: ['Reduces interference', 'Better performance'],
-            disadvantages: ['More expensive', 'Harder to install']
-          },
-          {
-            name: 'Coaxial Cable',
-            description: 'Cable with inner conductor surrounded by concentric shield',
-            uses: ['Cable TV', 'Internet connections', 'Video transmission']
-          },
-          {
-            name: 'Fiber Optic Cable',
-            description: 'Cable containing optical fibers for light transmission',
-            advantages: ['High speed', 'Long distance', 'No interference'],
-            disadvantages: ['Expensive', 'Fragile', 'Complex installation']
-          }
-        ],
-        colorCoding: {
-          t568a: {
-            name: 'T-568A Standard',
-            pairs: [
-              { pair: 1, colors: ['White/Green', 'Green'] },
-              { pair: 2, colors: ['White/Orange', 'Blue'] },
-              { pair: 3, colors: ['White/Blue', 'Orange'] },
-              { pair: 4, colors: ['White/Brown', 'Brown'] }
-            ]
-          },
-          t568b: {
-            name: 'T-568B Standard',
-            pairs: [
-              { pair: 1, colors: ['White/Orange', 'Orange'] },
-              { pair: 2, colors: ['White/Green', 'Blue'] },
-              { pair: 3, colors: ['White/Blue', 'Green'] },
-              { pair: 4, colors: ['White/Brown', 'Brown'] }
-            ]
-          }
-        },
-        crimpingSteps: [
-          'Strip cable jacket to expose 4 twisted pairs',
-          'Untwist pairs and arrange according to T-568A standard',
-          'Trim wires to uniform length',
-          'Insert wires into RJ45 connector',
-          'Crimp connector using crimping tool',
-          'Test cable with LAN tester'
-        ]
-      }
-    },
-    {
-      id: 'module3',
-      title: 'IP Addressing and Configuration',
-      description: 'IP addressing, subnetting, and network configuration',
-      duration: '10 hours',
-      videoUrl: 'https://www.youtube.com/watch?v=7_-qWlvQQtY',
-      topics: ['IPv4/IPv6', 'Static vs Dynamic IP', 'Subnetting', 'Network Configuration']
-    },
-    {
-      id: 'module4',
-      title: 'Router and Wireless Setup',
-      description: 'Configuring routers, Wi-Fi, and wireless access points',
-      duration: '8 hours',
-      videoUrl: 'https://www.youtube.com/watch?v=AMXIE5c5sv8',
-      topics: ['Router Configuration', 'Wi-Fi Setup', 'Access Points', 'Network Security'],
-      practiceContent: {
-        title: 'Router and Wireless Configuration',
-        type: 'B',
-        objectives: 'Configure routers and wireless access points for network connectivity',
-        materials: ['Router', 'Wireless access point', 'Computer system', 'Network cables'],
-        procedures: [
-          'Connect router to internet source',
-          'Access router configuration interface',
-          'Configure basic network settings',
-          'Set up wireless network',
-          'Configure security settings',
-          'Test wireless connectivity'
-        ],
-        routerSetup: [
-          'Connect router to modem using Ethernet cable',
-          'Power on router and wait for initialization',
-          'Access router admin interface (usually 192.168.1.1 or 192.168.0.1)',
-          'Login with default credentials (check router manual)',
-          'Configure WAN settings for internet connection',
-          'Set up LAN network settings',
-          'Configure DHCP server settings',
-          'Set up wireless network name (SSID)',
-          'Configure wireless security (WPA2/WPA3)',
-          'Save settings and restart router'
-        ],
-        wirelessSecurity: [
-          {
-            name: 'WPA2 (Wi-Fi Protected Access 2)',
-            description: 'Strong encryption for wireless networks',
-            advantages: ['High security', 'Widely supported'],
-            setup: 'Choose WPA2-PSK and set strong password'
-          },
-          {
-            name: 'WPA3 (Wi-Fi Protected Access 3)',
-            description: 'Latest wireless security standard',
-            advantages: ['Enhanced security', 'Future-proof'],
-            setup: 'Choose WPA3-SAE if supported by devices'
-          }
-        ]
-      }
-    }
   ]
 
   const networkTypes = [
@@ -431,755 +257,7 @@ export default function COC2Page() {
       </section>
 
       {/* Modules */}
-      <section className="py-16 bg-white">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="text-center mb-12">
-            <h2 className="text-3xl font-bold text-gray-900 mb-4">Course Modules</h2>
-            <p className="text-lg text-gray-600 max-w-3xl mx-auto">
-              Structured learning modules covering all aspects of network setup and configuration.
-            </p>
-          </div>
-          
-          <div className="space-y-6">
-            {modules.map((module, index) => (
-              <div key={module.id} className="bg-gray-50 rounded-lg p-6">
-                <div className="flex items-start justify-between">
-                  <div className="flex-1">
-                    <div className="flex items-center mb-3">
-                      <span className="bg-green-600 text-white w-8 h-8 rounded-full flex items-center justify-center text-sm font-bold mr-4">
-                        {index + 1}
-                      </span>
-                      <h3 className="text-xl font-semibold text-gray-900">{module.title}</h3>
-                    </div>
-                    <p className="text-gray-600 mb-4">{module.description}</p>
-                    <div className="flex items-center text-sm text-gray-500 mb-4">
-                      <Clock className="h-4 w-4 mr-1" />
-                      {module.duration}
-                    </div>
-                    <div className="grid md:grid-cols-2 gap-2">
-                      {module.topics.map((topic, topicIndex) => (
-                        <div key={topicIndex} className="flex items-center text-sm text-gray-600">
-                          <ChevronRight className="h-3 w-3 mr-2 text-green-500" />
-                          {topic}
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                  <div className="ml-6 flex space-x-2">
-                    <button 
-                      onClick={() => {
-                        const modal = document.getElementById(`video-modal-${module.id}`);
-                        if (modal) modal.classList.remove('hidden');
-                      }}
-                      className="bg-green-600 text-white px-4 py-2 rounded-lg hover:bg-green-700 transition-colors flex items-center"
-                    >
-                      <Video className="h-4 w-4 mr-2" />
-                      Watch
-                    </button>
-                    <button 
-                      onClick={() => {
-                        const modal = document.getElementById(`practice-modal-${module.id}`);
-                        if (modal) modal.classList.remove('hidden');
-                      }}
-                      className="bg-blue-600 text-white px-4 py-2 rounded-lg hover:bg-blue-700 transition-colors flex items-center"
-                    >
-                      <FileText className="h-4 w-4 mr-2" />
-                      Practice
-                    </button>
-                  </div>
-                </div>
-
-                {/* Video Modal */}
-                <div id={`video-modal-${module.id}`} className="hidden fixed inset-0 bg-black bg-opacity-50 z-50 flex items-center justify-center p-4">
-                  <div className="bg-white rounded-lg max-w-4xl w-full max-h-[90vh] overflow-y-auto">
-                    <div className="p-6">
-                      <div className="flex justify-between items-center mb-6">
-                        <h3 className="text-2xl font-bold text-gray-900">{module.title}</h3>
-                        <button 
-                          onClick={() => {
-                            const modal = document.getElementById(`video-modal-${module.id}`);
-                            const iframe = modal?.querySelector('iframe');
-                              if (iframe) {
-                                // Stop the video by changing the src
-                                const currentSrc = iframe.src;
-                                iframe.src = '';
-                                iframe.src = currentSrc;
-                              }
-                              if (modal) modal.classList.add('hidden');
-                          }}
-                          className="text-gray-500 hover:text-gray-700"
-                        >
-                          <svg className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                          </svg>
-                        </button>
-                      </div>
-
-                      <div className="aspect-video w-full">
-                        <iframe
-                          src={module.videoUrl.replace('watch?v=', 'embed/')}
-                          title={`${module.title} Video`}
-                          className="w-full h-full rounded-lg"
-                          frameBorder="0"
-                          allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-                          allowFullScreen
-                        ></iframe>
-                      </div>
-
-                      <div className="mt-6">
-                        <h4 className="text-lg font-semibold text-gray-900 mb-2">Module Description</h4>
-                        <p className="text-gray-600">{module.description}</p>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-
-                {/* Practice Modal */}
-                <div id={`practice-modal-${module.id}`} className="hidden fixed inset-0 bg-black bg-opacity-50 z-50 flex items-center justify-center p-4">
-                  <div className="bg-white rounded-lg max-w-4xl w-full max-h-[90vh] overflow-y-auto">
-                    <div className="p-6">
-                      <div className="flex justify-between items-center mb-6">
-                        <h3 className="text-2xl font-bold text-gray-900">{module.title}</h3>
-                        <button 
-                          onClick={() => {
-                            const modal = document.getElementById(`practice-modal-${module.id}`);
-                            if (modal) modal.classList.add('hidden');
-                          }}
-                          className="text-gray-500 hover:text-gray-700"
-                        >
-                          <svg className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                          </svg>
-                        </button>
-                      </div>
-
-                      {/* Module 1 Specific Content */}
-                      {module.id === 'module1' && (
-                        <div className="space-y-6">
-                          <div className="bg-green-50 rounded-lg p-4">
-                            <h4 className="font-semibold text-green-900 mb-2">Objectives:</h4>
-                            <p className="text-green-800">Understand different network types and topology configurations</p>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Materials Needed:</h4>
-                            <ul className="list-disc list-inside space-y-1 text-gray-700">
-                              <li>Network diagrams</li>
-                              <li>Computer systems</li>
-                              <li>Network simulation software</li>
-                            </ul>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Step-by-Step Procedures:</h4>
-                            <ol className="list-decimal list-inside space-y-2 text-gray-700">
-                              <li>Study different network types (LAN, WAN, PAN, MAN)</li>
-                              <li>Analyze network topology diagrams</li>
-                              <li>Identify network components and their functions</li>
-                              <li>Practice network design principles</li>
-                              <li>Document network architecture</li>
-                            </ol>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Network Types:</h4>
-                            <div className="grid md:grid-cols-2 gap-4">
-                              <div className="bg-gray-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-gray-900 mb-2">Local Area Networks (LAN)</h5>
-                                <p className="text-sm text-gray-600 mb-2">Network that connects computers within a limited area</p>
-                                <div>
-                                  <span className="text-xs font-medium text-gray-500">Examples:</span>
-                                  <ul className="text-xs text-gray-600 mt-1">
-                                    <li>• School network</li>
-                                    <li>• Office network</li>
-                                    <li>• Home network</li>
-                                  </ul>
-                                </div>
-                              </div>
-                              <div className="bg-gray-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-gray-900 mb-2">Wide Area Networks (WAN)</h5>
-                                <p className="text-sm text-gray-600 mb-2">Network that covers larger geographical areas</p>
-                                <div>
-                                  <span className="text-xs font-medium text-gray-500">Examples:</span>
-                                  <ul className="text-xs text-gray-600 mt-1">
-                                    <li>• Internet</li>
-                                    <li>• Corporate network across cities</li>
-                                  </ul>
-                                </div>
-                              </div>
-                              <div className="bg-gray-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-gray-900 mb-2">Personal Area Networks (PAN)</h5>
-                                <p className="text-sm text-gray-600 mb-2">Network based on individual workspace</p>
-                                <div>
-                                  <span className="text-xs font-medium text-gray-500">Examples:</span>
-                                  <ul className="text-xs text-gray-600 mt-1">
-                                    <li>• Bluetooth devices</li>
-                                    <li>• Personal computer setup</li>
-                                  </ul>
-                                </div>
-                              </div>
-                              <div className="bg-gray-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-gray-900 mb-2">Metropolitan Area Networks (MAN)</h5>
-                                <p className="text-sm text-gray-600 mb-2">Network stretching across metropolitan areas</p>
-                                <div>
-                                  <span className="text-xs font-medium text-gray-500">Examples:</span>
-                                  <ul className="text-xs text-gray-600 mt-1">
-                                    <li>• City-wide network</li>
-                                    <li>• University campus network</li>
-                                  </ul>
-                                </div>
-                              </div>
-                            </div>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Network Topology Types:</h4>
-                            <div className="grid md:grid-cols-2 gap-4">
-                              <div className="bg-blue-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-blue-900 mb-2">Star Topology</h5>
-                                <p className="text-sm text-blue-800 mb-3">Central node connects to each computer independently</p>
-                                <div className="grid grid-cols-2 gap-3">
-                                  <div>
-                                    <span className="text-xs font-medium text-blue-700">Advantages:</span>
-                                    <ul className="text-xs text-blue-600 mt-1">
-                                      <li>• Easy to manage</li>
-                                      <li>• One failure doesn&apos;t affect others</li>
-                                    </ul>
-                                  </div>
-                                  <div>
-                                    <span className="text-xs font-medium text-blue-700">Disadvantages:</span>
-                                    <ul className="text-xs text-blue-600 mt-1">
-                                      <li>• Requires many cables</li>
-                                      <li>• Central point failure</li>
-                                    </ul>
-                                  </div>
-                                </div>
-                              </div>
-                              <div className="bg-blue-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-blue-900 mb-2">Bus Topology</h5>
-                                <p className="text-sm text-blue-800 mb-3">Single cable connects all computers</p>
-                                <div className="grid grid-cols-2 gap-3">
-                                  <div>
-                                    <span className="text-xs font-medium text-blue-700">Advantages:</span>
-                                    <ul className="text-xs text-blue-600 mt-1">
-                                      <li>• Less cabling required</li>
-                                      <li>• Simple setup</li>
-                                    </ul>
-                                  </div>
-                                  <div>
-                                    <span className="text-xs font-medium text-blue-700">Disadvantages:</span>
-                                    <ul className="text-xs text-blue-600 mt-1">
-                                      <li>• Single point of failure</li>
-                                      <li>• Difficult troubleshooting</li>
-                                    </ul>
-                                  </div>
-                                </div>
-                              </div>
-                              <div className="bg-blue-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-blue-900 mb-2">Ring Topology</h5>
-                                <p className="text-sm text-blue-800 mb-3">Computers connected in a circular pattern</p>
-                                <div className="grid grid-cols-2 gap-3">
-                                  <div>
-                                    <span className="text-xs font-medium text-blue-700">Advantages:</span>
-                                    <ul className="text-xs text-blue-600 mt-1">
-                                      <li>• Equal access to network</li>
-                                      <li>• Good performance</li>
-                                    </ul>
-                                  </div>
-                                  <div>
-                                    <span className="text-xs font-medium text-blue-700">Disadvantages:</span>
-                                    <ul className="text-xs text-blue-600 mt-1">
-                                      <li>• Single failure affects all</li>
-                                      <li>• Difficult to expand</li>
-                                    </ul>
-                                  </div>
-                                </div>
-                              </div>
-                              <div className="bg-blue-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-blue-900 mb-2">Mesh Topology</h5>
-                                <p className="text-sm text-blue-800 mb-3">All nodes interconnected with each other</p>
-                                <div className="grid grid-cols-2 gap-3">
-                                  <div>
-                                    <span className="text-xs font-medium text-blue-700">Advantages:</span>
-                                    <ul className="text-xs text-blue-600 mt-1">
-                                      <li>• Highly reliable</li>
-                                      <li>• Multiple paths available</li>
-                                    </ul>
-                                  </div>
-                                  <div>
-                                    <span className="text-xs font-medium text-blue-700">Disadvantages:</span>
-                                    <ul className="text-xs text-blue-600 mt-1">
-                                      <li>• Expensive</li>
-                                      <li>• Complex configuration</li>
-                                    </ul>
-                                  </div>
-                                </div>
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-                      )}
-
-                      {/* Module 2 Specific Content */}
-                      {module.id === 'module2' && (
-                        <div className="space-y-6">
-                          <div className="bg-green-50 rounded-lg p-4">
-                            <h4 className="font-semibold text-green-900 mb-2">Objectives:</h4>
-                            <p className="text-green-800">Create cable connection using UTP cable with specific standard tools and materials in accordance to cable color coding</p>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Materials Needed:</h4>
-                            <ul className="list-disc list-inside space-y-1 text-gray-700">
-                              <li>UTP Cable</li>
-                              <li>RJ45 Connectors</li>
-                              <li>Crimping Tools</li>
-                              <li>LAN Cable Tester</li>
-                            </ul>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Step-by-Step Procedures:</h4>
-                            <ol className="list-decimal list-inside space-y-2 text-gray-700">
-                              <li>Get a UTP cable and crimp both ends using crimping tools and RJ45</li>
-                              <li>Perform color coding T-568A</li>
-                              <li>Test the cable using LAN cable tester</li>
-                            </ol>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Cable Types:</h4>
-                            <div className="grid md:grid-cols-2 gap-4">
-                              <div className="bg-gray-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-gray-900 mb-2">Unshielded Twisted Pair (UTP)</h5>
-                                <p className="text-sm text-gray-600 mb-2">Copper cabling with twisted pairs without shielding</p>
-                                <div className="mb-2">
-                                  <span className="text-xs font-medium text-gray-500">Categories:</span>
-                                  <p className="text-xs text-gray-600 mt-1">Cat5e, Cat6, Cat6a, Cat7</p>
-                                </div>
-                                <div>
-                                  <span className="text-xs font-medium text-gray-500">Uses:</span>
-                                  <ul className="text-xs text-gray-600 mt-1">
-                                    <li>• Ethernet networks</li>
-                                    <li>• Telephone systems</li>
-                                    <li>• Data transmission</li>
-                                  </ul>
-                                </div>
-                              </div>
-                              <div className="bg-gray-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-gray-900 mb-2">Shielded Twisted Pair (STP)</h5>
-                                <p className="text-sm text-gray-600 mb-2">Copper cabling with additional shielding</p>
-                                <div className="mb-2">
-                                  <span className="text-xs font-medium text-gray-500">Advantages:</span>
-                                  <ul className="text-xs text-gray-600 mt-1">
-                                    <li>• Reduces interference</li>
-                                    <li>• Better performance</li>
-                                  </ul>
-                                </div>
-                                <div>
-                                  <span className="text-xs font-medium text-gray-500">Disadvantages:</span>
-                                  <ul className="text-xs text-gray-600 mt-1">
-                                    <li>• More expensive</li>
-                                    <li>• Harder to install</li>
-                                  </ul>
-                                </div>
-                              </div>
-                              <div className="bg-gray-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-gray-900 mb-2">Coaxial Cable</h5>
-                                <p className="text-sm text-gray-600 mb-2">Cable with inner conductor surrounded by concentric shield</p>
-                                <div>
-                                  <span className="text-xs font-medium text-gray-500">Uses:</span>
-                                  <ul className="text-xs text-gray-600 mt-1">
-                                    <li>• Cable TV</li>
-                                    <li>• Internet connections</li>
-                                    <li>• Video transmission</li>
-                                  </ul>
-                                </div>
-                              </div>
-                              <div className="bg-gray-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-gray-900 mb-2">Fiber Optic Cable</h5>
-                                <p className="text-sm text-gray-600 mb-2">Cable containing optical fibers for light transmission</p>
-                                <div className="mb-2">
-                                  <span className="text-xs font-medium text-gray-500">Advantages:</span>
-                                  <ul className="text-xs text-gray-600 mt-1">
-                                    <li>• High speed</li>
-                                    <li>• Long distance</li>
-                                    <li>• No interference</li>
-                                  </ul>
-                                </div>
-                                <div>
-                                  <span className="text-xs font-medium text-gray-500">Disadvantages:</span>
-                                  <ul className="text-xs text-gray-600 mt-1">
-                                    <li>• Expensive</li>
-                                    <li>• Fragile</li>
-                                    <li>• Complex installation</li>
-                                  </ul>
-                                </div>
-                              </div>
-                            </div>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Color Coding Standards:</h4>
-                            <div className="grid md:grid-cols-2 gap-6">
-                              <div className="bg-yellow-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-yellow-900 mb-3">T-568A Standard</h5>
-                                <div className="space-y-2">
-                                  <div className="flex items-center text-sm">
-                                    <span className="font-medium text-yellow-800 mr-2">Pair 1:</span>
-                                    <span className="text-yellow-700">White/Green - Green</span>
-                                  </div>
-                                  <div className="flex items-center text-sm">
-                                    <span className="font-medium text-yellow-800 mr-2">Pair 2:</span>
-                                    <span className="text-yellow-700">White/Orange - Blue</span>
-                                  </div>
-                                  <div className="flex items-center text-sm">
-                                    <span className="font-medium text-yellow-800 mr-2">Pair 3:</span>
-                                    <span className="text-yellow-700">White/Blue - Orange</span>
-                                  </div>
-                                  <div className="flex items-center text-sm">
-                                    <span className="font-medium text-yellow-800 mr-2">Pair 4:</span>
-                                    <span className="text-yellow-700">White/Brown - Brown</span>
-                                  </div>
-                                </div>
-                              </div>
-                              <div className="bg-blue-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-blue-900 mb-3">T-568B Standard</h5>
-                                <div className="space-y-2">
-                                  <div className="flex items-center text-sm">
-                                    <span className="font-medium text-blue-800 mr-2">Pair 1:</span>
-                                    <span className="text-blue-700">White/Orange - Orange</span>
-                                  </div>
-                                  <div className="flex items-center text-sm">
-                                    <span className="font-medium text-blue-800 mr-2">Pair 2:</span>
-                                    <span className="text-blue-700">White/Green - Blue</span>
-                                  </div>
-                                  <div className="flex items-center text-sm">
-                                    <span className="font-medium text-blue-800 mr-2">Pair 3:</span>
-                                    <span className="text-blue-700">White/Blue - Green</span>
-                                  </div>
-                                  <div className="flex items-center text-sm">
-                                    <span className="font-medium text-blue-800 mr-2">Pair 4:</span>
-                                    <span className="text-blue-700">White/Brown - Brown</span>
-                                  </div>
-                                </div>
-                              </div>
-                            </div>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Crimping Steps:</h4>
-                            <ol className="list-decimal list-inside space-y-2 text-gray-700">
-                              <li>Strip cable jacket to expose 4 twisted pairs</li>
-                              <li>Untwist pairs and arrange according to T-568A standard</li>
-                              <li>Trim wires to uniform length</li>
-                              <li>Insert wires into RJ45 connector</li>
-                              <li>Crimp connector using crimping tool</li>
-                              <li>Test cable with LAN tester</li>
-                            </ol>
-                          </div>
-                        </div>
-                      )}
-
-                      {/* Module 3 Specific Content */}
-                      {module.id === 'module3' && (
-                        <div className="space-y-6">
-                          <div className="bg-green-50 rounded-lg p-4">
-                            <h4 className="font-semibold text-green-900 mb-2">Objectives:</h4>
-                            <p className="text-green-800">Set up network configuration and configure IP addressing for network devices</p>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Materials Needed:</h4>
-                            <ul className="list-disc list-inside space-y-1 text-gray-700">
-                              <li>Computer system</li>
-                              <li>Network interface card</li>
-                              <li>Router/switch</li>
-                              <li>Network cables</li>
-                            </ul>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Step-by-Step Procedures:</h4>
-                            <ol className="list-decimal list-inside space-y-2 text-gray-700">
-                              <li>Configure network interface settings</li>
-                              <li>Set up static IP addressing</li>
-                              <li>Configure subnet mask and gateway</li>
-                              <li>Test network connectivity</li>
-                              <li>Document network configuration</li>
-                            </ol>
-                          </div>
-
-                          {/* Static Configuration */}
-                          <div className="bg-blue-50 rounded-lg p-4">
-                            <h4 className="font-semibold text-blue-900 mb-2">Static IP Configuration Setup</h4>
-                            <p className="text-blue-800 mb-3">Manual IP address assignment for network devices</p>
-                            
-                            <div className="mb-4">
-                              <h5 className="font-medium text-blue-900 mb-2">Network Setup:</h5>
-                              <div className="grid md:grid-cols-2 gap-3 text-sm">
-                                <div className="bg-white rounded p-2">
-                                  <span className="font-medium text-blue-800">Server:</span>
-                                  <div className="text-blue-700">IP: 10.0.0.100</div>
-                                  <div className="text-blue-700">Subnet: 255.255.255.0</div>
-                                  <div className="text-blue-700">Gateway: 10.0.0.1</div>
-                                </div>
-                                <div className="bg-white rounded p-2">
-                                  <span className="font-medium text-blue-800">Laptop:</span>
-                                  <div className="text-blue-700">IP: 10.0.0.101</div>
-                                  <div className="text-blue-700">Subnet: 255.255.255.0</div>
-                                  <div className="text-blue-700">Gateway: 10.0.0.1</div>
-                                </div>
-                                <div className="bg-white rounded p-2">
-                                  <span className="font-medium text-blue-800">Desktop 1:</span>
-                                  <div className="text-blue-700">IP: 10.0.0.102</div>
-                                  <div className="text-blue-700">Subnet: 255.255.255.0</div>
-                                  <div className="text-blue-700">Gateway: 10.0.0.1</div>
-                                </div>
-                                <div className="bg-white rounded p-2">
-                                  <span className="font-medium text-blue-800">Desktop 2:</span>
-                                  <div className="text-blue-700">IP: 10.0.0.103</div>
-                                  <div className="text-blue-700">Subnet: 255.255.255.0</div>
-                                  <div className="text-blue-700">Gateway: 10.0.0.1</div>
-                                </div>
-                              </div>
-                            </div>
-
-                            <div className="space-y-4">
-                              <div>
-                                <h6 className="font-medium text-blue-900 mb-2">IP Configuration Steps:</h6>
-                                <ol className="list-decimal list-inside space-y-1 text-sm text-blue-700">
-                                  <li>Press Windows + R, type &quot;ncpa.cpl&quot; and press enter</li>
-                                  <li>Right-click Ethernet connection and select Properties</li>
-                                  <li>Select &quot;Internet Protocol Version 4 (TCP/IPv4)&quot; and click Properties</li>
-                                  <li>Check &quot;Use the following IP Address&quot;</li>
-                                  <li>Enter the assigned IP address for your device</li>
-                                  <li>Enter Subnet Mask: 255.255.255.0</li>
-                                  <li>Enter Default Gateway: 10.0.0.1</li>
-                                  <li>Click OK to save settings</li>
-                                </ol>
-                              </div>
-
-                              <div>
-                                <h6 className="font-medium text-blue-900 mb-2">Firewall Configuration:</h6>
-                                <ol className="list-decimal list-inside space-y-1 text-sm text-blue-700">
-                                  <li>Press Windows + R, type &quot;control firewall.cpl&quot; and press enter</li>
-                                  <li>Turn OFF Windows Defender Firewall for Private, Public, and Domain networks</li>
-                                  <li>Click OK to save changes</li>
-                                </ol>
-                              </div>
-
-                              <div>
-                                <h6 className="font-medium text-blue-900 mb-2">Network Sharing Setup:</h6>
-                                <ol className="list-decimal list-inside space-y-1 text-sm text-blue-700">
-                                  <li>Press Windows + R, type &quot;control.exe /name Microsoft.NetworkAndSharingCenter&quot;</li>
-                                  <li>Click &quot;Change advanced sharing settings&quot;</li>
-                                  <li>Turn ON Network Discovery</li>
-                                  <li>Turn ON File and Printer Sharing</li>
-                                  <li>Turn OFF Password Protected Sharing</li>
-                                  <li>Click &quot;Save Changes&quot;</li>
-                                </ol>
-                              </div>
-
-                              <div>
-                                <h6 className="font-medium text-blue-900 mb-2">Required Services:</h6>
-                                <ol className="list-decimal list-inside space-y-1 text-sm text-blue-700">
-                                  <li>Press Windows + R, type &quot;services.msc&quot; and press enter</li>
-                                  <li>Right-click each service, set Startup Type to Automatic and Start the service:</li>
-                                  <li>• Functional Discovery Resource Publication</li>
-                                  <li>• SSDP Discovery</li>
-                                  <li>• UPnP Device Host</li>
-                                </ol>
-                              </div>
-
-                              <div>
-                                <h6 className="font-medium text-blue-900 mb-2">Connectivity Testing:</h6>
-                                <ol className="list-decimal list-inside space-y-1 text-sm text-blue-700">
-                                  <li>Press Windows + R, type &quot;cmd&quot; and press enter</li>
-                                  <li>Type &quot;ping [IP_ADDRESS] -t&quot; to test connectivity</li>
-                                  <li>Example: ping 10.0.0.101 -t</li>
-                                </ol>
-                              </div>
-                            </div>
-                          </div>
-
-                          {/* Dynamic Configuration */}
-                          <div className="bg-green-50 rounded-lg p-4">
-                            <h4 className="font-semibold text-green-900 mb-2">Dynamic IP Configuration Setup</h4>
-                            <p className="text-green-800 mb-3">Automatic IP address assignment using DHCP</p>
-                            
-                            <div className="space-y-4">
-                              <div>
-                                <h6 className="font-medium text-green-900 mb-2">Router Configuration:</h6>
-                                <ol className="list-decimal list-inside space-y-1 text-sm text-green-700">
-                                  <li>Turn ON network router and connect via LAN cable</li>
-                                  <li>Hold RESET button for 5 seconds until power LED blinks</li>
-                                  <li>Check gateway IP: Press Windows + R, type &quot;ncpa.cpl&quot;</li>
-                                  <li>Right-click Ethernet, click Status, click Details</li>
-                                  <li>Look for IPv4 Default Gateway (e.g., 192.168.0.1)</li>
-                                  <li>Open web browser and type the gateway IP</li>
-                                  <li>Login with default credentials (admin/admin)</li>
-                                </ol>
-                              </div>
-
-                              <div>
-                                <h6 className="font-medium text-green-900 mb-2">Wireless Settings:</h6>
-                                <ol className="list-decimal list-inside space-y-1 text-sm text-green-700">
-                                  <li>Go to Wireless Menu</li>
-                                  <li>Change Wireless Network Name to Lastname.Firstname</li>
-                                  <li>Go to Wireless Security</li>
-                                  <li>Under WPA/WPA2, change Wireless Password to 123456789</li>
-                                </ol>
-                              </div>
-
-                              <div>
-                                <h6 className="font-medium text-green-900 mb-2">Client Configuration:</h6>
-                                <ol className="list-decimal list-inside space-y-1 text-sm text-green-700">
-                                  <li>Press Windows + R, type &quot;ncpa.cpl&quot; and press enter</li>
-                                  <li>Right-click Ethernet connection and select Properties</li>
-                                  <li>Select &quot;Internet Protocol Version 4 (TCP/IPv4)&quot; and click Properties</li>
-                                  <li>Check &quot;Obtain an IP address automatically&quot;</li>
-                                  <li>Check &quot;Obtain DNS server address automatically&quot;</li>
-                                  <li>Click OK to save settings</li>
-                                </ol>
-                              </div>
-
-                              <div>
-                                <h6 className="font-medium text-green-900 mb-2">Firewall Configuration:</h6>
-                                <ol className="list-decimal list-inside space-y-1 text-sm text-green-700">
-                                  <li>Press Windows + R, type &quot;control firewall.cpl&quot; and press enter</li>
-                                  <li>Turn OFF Windows Defender Firewall for Private, Public, and Domain networks</li>
-                                  <li>Click OK to save changes</li>
-                                </ol>
-                              </div>
-
-                              <div>
-                                <h6 className="font-medium text-green-900 mb-2">Network Sharing Setup:</h6>
-                                <ol className="list-decimal list-inside space-y-1 text-sm text-green-700">
-                                  <li>Press Windows + R, type &quot;control.exe /name Microsoft.NetworkAndSharingCenter&quot;</li>
-                                  <li>Click &quot;Change advanced sharing settings&quot;</li>
-                                  <li>Turn ON Network Discovery</li>
-                                  <li>Turn ON File and Printer Sharing</li>
-                                  <li>Turn OFF Password Protected Sharing</li>
-                                  <li>Click &quot;Save Changes&quot;</li>
-                                </ol>
-                              </div>
-
-                              <div>
-                                <h6 className="font-medium text-green-900 mb-2">Required Services:</h6>
-                                <ol className="list-decimal list-inside space-y-1 text-sm text-green-700">
-                                  <li>Press Windows + R, type &quot;services.msc&quot; and press enter</li>
-                                  <li>Right-click each service, set Startup Type to Automatic and Start the service:</li>
-                                  <li>• Functional Discovery Resource Publication</li>
-                                  <li>• SSDP Discovery</li>
-                                  <li>• UPnP Device Host</li>
-                                </ol>
-                              </div>
-
-                              <div>
-                                <h6 className="font-medium text-green-900 mb-2">Connectivity Testing:</h6>
-                                <ol className="list-decimal list-inside space-y-1 text-sm text-green-700">
-                                  <li>Press Windows + R, type &quot;cmd&quot; and press enter</li>
-                                  <li>Type &quot;ping [IP_ADDRESS] -t&quot; to test connectivity</li>
-                                  <li>Example: ping 192.168.0.101 -t</li>
-                                </ol>
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-                      )}
-
-                      {/* Module 4 Specific Content */}
-                      {module.id === 'module4' && (
-                        <div className="space-y-6">
-                          <div className="bg-green-50 rounded-lg p-4">
-                            <h4 className="font-semibold text-green-900 mb-2">Objectives:</h4>
-                            <p className="text-green-800">Configure routers and wireless access points for network connectivity</p>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Materials Needed:</h4>
-                            <ul className="list-disc list-inside space-y-1 text-gray-700">
-                              <li>Router</li>
-                              <li>Wireless access point</li>
-                              <li>Computer system</li>
-                              <li>Network cables</li>
-                            </ul>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Step-by-Step Procedures:</h4>
-                            <ol className="list-decimal list-inside space-y-2 text-gray-700">
-                              <li>Connect router to internet source</li>
-                              <li>Access router configuration interface</li>
-                              <li>Configure basic network settings</li>
-                              <li>Set up wireless network</li>
-                              <li>Configure security settings</li>
-                              <li>Test wireless connectivity</li>
-                            </ol>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Router Setup Steps:</h4>
-                            <ol className="list-decimal list-inside space-y-2 text-gray-700">
-                              <li>Connect router to modem using Ethernet cable</li>
-                              <li>Power on router and wait for initialization</li>
-                              <li>Access router admin interface (usually 192.168.1.1 or 192.168.0.1)</li>
-                              <li>Login with default credentials (check router manual)</li>
-                              <li>Configure WAN settings for internet connection</li>
-                              <li>Set up LAN network settings</li>
-                              <li>Configure DHCP server settings</li>
-                              <li>Set up wireless network name (SSID)</li>
-                              <li>Configure wireless security (WPA2/WPA3)</li>
-                              <li>Save settings and restart router</li>
-                            </ol>
-                          </div>
-
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Wireless Security:</h4>
-                            <div className="grid md:grid-cols-2 gap-4">
-                              <div className="bg-green-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-green-900 mb-2">WPA2 (Wi-Fi Protected Access 2)</h5>
-                                <p className="text-sm text-green-800 mb-2">Strong encryption for wireless networks</p>
-                                <div className="mb-2">
-                                  <span className="text-xs font-medium text-green-700">Advantages:</span>
-                                  <ul className="text-xs text-green-600 mt-1">
-                                    <li>• High security</li>
-                                    <li>• Widely supported</li>
-                                  </ul>
-                                </div>
-                                <div>
-                                  <span className="text-xs font-medium text-green-700">Setup:</span>
-                                  <p className="text-xs text-green-600 mt-1">Choose WPA2-PSK and set strong password</p>
-                                </div>
-                              </div>
-                              <div className="bg-green-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-green-900 mb-2">WPA3 (Wi-Fi Protected Access 3)</h5>
-                                <p className="text-sm text-green-800 mb-2">Latest wireless security standard</p>
-                                <div className="mb-2">
-                                  <span className="text-xs font-medium text-green-700">Advantages:</span>
-                                  <ul className="text-xs text-green-600 mt-1">
-                                    <li>• Enhanced security</li>
-                                    <li>• Future-proof</li>
-                                  </ul>
-                                </div>
-                                <div>
-                                  <span className="text-xs font-medium text-green-700">Setup:</span>
-                                  <p className="text-xs text-green-600 mt-1">Choose WPA3-SAE if supported by devices</p>
-                                </div>
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-                      )}
-                    </div>
-                  </div>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      </section>
+      <COC2Modules />
 
       {/* Key Topics */}
       <section className="py-16 bg-gray-50">
