+++ src/app/coc3/page.tsx
@@ -1,322 +1,140 @@
-"use client"
-
 import Link from 'next/link'
+import { Metadata } from 'next'
 import { 
   Server, 
   ArrowLeft, 
   BookOpen, 
-  Video, 
-  FileText, 
   CheckCircle,
   Clock,
   Users,
   ChevronRight,
   Shield,
-  Database
+  Database,
+  Settings,
+  Monitor,
+  FileText
 } from 'lucide-react'
 import Footer from '@/components/Footer'
+import COC3Modules from './COC3Modules'
+
+export const metadata: Metadata = {
+  title: 'COC 3: Configure Server - CSS NC II Training',
+  description: 'Learn server configuration and management including local server setup, Active Directory installation, folder redirection, and remote desktop management. Master domain controller setup and network services.',
+  keywords: 'COC 3, Server Configuration, Active Directory, Domain Controller, Folder Redirection, Remote Desktop, Network Services, Server Management, CSS NC II, TESDA',
+  openGraph: {
+    title: 'COC 3: Configure Server - CSS NC II Training',
+    description: 'Learn server configuration and management including local server setup, Active Directory installation, folder redirection, and remote desktop management.',
+    url: 'https://css-nc2-companion.vercel.app/coc3',
+    siteName: 'CSS NC II Training Companion App',
+    images: [
+      {
+        url: '/img/meta-coc3.jpg',
+        width: 1200,
+        height: 630,
+        alt: 'COC 3: Configure Server',
+      },
+    ],
+    locale: 'en_US',
+    type: 'website',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'COC 3: Configure Server - CSS NC II Training',
+    description: 'Learn server configuration and management including local server setup, Active Directory installation, folder redirection, and remote desktop management.',
+    images: ['/img/meta-coc3.jpg'],
+  },
+};
 
 export default function COC3Page() {
   const learningObjectives = [
-    'Set up and configure computer servers',
-    'Configure user access and permissions',
-    'Set up network services and protocols',
-    'Implement server security measures',
-    'Perform server testing and validation',
-    'Create and maintain documentation'
-  ]
-
-  const modules = [
-    {
-      id: 'module1',
-      title: 'Server Fundamentals',
-      description: 'Understanding server types, roles, and basic administration',
-      duration: '8 hours',
-      topics: ['Server Types', 'Server Roles', 'Basic Administration', 'Server Hardware'],
-      practiceContent: {
-        title: 'Local Server Configuration',
-        type: 'A',
-        objectives: 'Configure local server with proper network settings, firewall configuration, and basic server setup',
-        materials: ['Windows Server', 'Network cables', 'Router/switch', 'Computer system'],
-        procedures: [
-          'Disable DHCP Server of the AP router first',
-          'Configure firewall settings',
-          'Enable Remote Desktop Connection',
-          'Setup Static IP Configuration',
-          'Rename computer to "Server" and restart'
-        ],
-        firewallSettings: {
-          title: 'Firewall Configuration',
-          description: 'Allow applications and features through Windows Firewall',
-          services: [
-            'Netlogon Service',
-            'Remote Desktop',
-            'Remote Service Management',
-            'Windows Firewall Remote Management'
-          ]
-        },
-        networkConfiguration: {
-          title: 'Static IP Configuration',
-          settings: {
-            'IP Address': '10.0.0.5',
-            'Subnet Mask': '255.255.255.0',
-            'Gateway': '10.0.0.1',
-            'Preferred DNS': '10.0.0.5'
-          }
-        }
-      }
-    },
-    {
-      id: 'module2',
-      title: 'User Access Management',
-      description: 'Setting up user accounts, groups, and access permissions',
-      duration: '10 hours',
-      topics: ['User Accounts', 'Group Management', 'Permissions', 'Access Control'],
-      practiceContent: {
-        title: 'Active Directory and User Management',
-        type: 'B',
-        objectives: 'Install Active Directory Domain Services, configure DHCP server, and create organizational users',
-        materials: ['Windows Server', 'Domain controller', 'DHCP server', 'Active Directory tools'],
-        procedures: [
-          'Install Active Directory Domain Services',
-          'Configure DNS Server',
-          'Promote server to Domain Controller',
-          'Install and configure DHCP Server',
-          'Create organizational units and user accounts'
-        ],
-        activeDirectorySetup: {
-          title: 'Active Directory Domain Services Installation',
-          steps: [
-            'Click Add roles and features',
-            'Select Role-based or feature-based installation',
-            'Check and select server or virtual hard disk',
-            'Select Active Directory Domain Services and DNS Server',
-            'Click Add Features on all roles or features',
-            'Click promote on Active Directory Domain Services notification',
-            'Select Add New Forest',
-            'Set root domain name to "bgfccss.com"',
-            'Set password to "Admin@12345"',
-            'Click Install'
-          ]
-        },
-        dhcpConfiguration: {
-          title: 'DHCP Server Configuration',
-          steps: [
-            'Go to Tools, select DHCP',
-            'Select and expand IPv4',
-            'Right click and select New Scope',
-            'Provide scope name: "LAN Scope"',
-            'Set IP range: 10.0.0.100 - 10.0.0.120',
-            'Set Length: 24',
-            'Set Subnet Mask: 255.255.255.0'
-          ]
-        },
-        userCreation: {
-          title: 'Organizational User Creation',
-          steps: [
-            'Go to Tools, select Active Directory Users and Computers',
-            'Right click on "bgfccss.com", select New, click Organizational Unit',
-            'Name it "Redirection OU"',
-            'Under "Redirection OU", create two User accounts:',
-            'Username: win10, Password: Admin@123456789',
-            'Username: win11, Password: Admin@123456789'
-          ]
-        }
-      }
-    },
-    {
-      id: 'module3',
-      title: 'Network Services Configuration',
-      description: 'Configuring essential network services and protocols',
-      duration: '12 hours',
-      topics: ['DHCP Services', 'DNS Configuration', 'File Services', 'Print Services'],
-      practiceContent: {
-        title: 'Group Policy and Folder Redirection',
-        type: 'B',
-        objectives: 'Configure Group Policy Management, create shared folders, and set up folder redirection',
-        materials: ['Windows Server', 'Group Policy Management', 'Shared folders', 'Network path'],
-        procedures: [
-          'Create folder and configure sharing permissions',
-          'Configure Group Policy Object',
-          'Set up folder redirection',
-          'Update group policy',
-          'Test folder redirection configuration'
-        ],
-        folderSharing: {
-          title: 'Folder Sharing and Permissions',
-          steps: [
-            'On Drive D or DATA, create folder named "Redirection"',
-            'Right click folder, select properties, select sharing tab',
-            'Click advance sharing, check share this folder',
-            'Click permission, under Everyone account, check Full Control',
-            'Click apply and OK',
-            'Copy Network Path (e.g., \\Server\\FolderName) to notepad'
-          ]
-        },
-        groupPolicy: {
-          title: 'Group Policy Object Configuration',
-          steps: [
-            'Go to Tools, Select Group Policy Management',
-            'Expand forest: bgfccss.com',
-            'Right click on "Redirection OU", select Create a GPO in this domain',
-            'Provide name: "Redirection GPO"',
-            'Right click the GPO, select Edit',
-            'Select User Configuration, Policies, Windows Settings, Folder Redirection',
-            'Select Desktop, right click, select properties',
-            'Select Basic-Redirect everyone\'s folder to the same location',
-            'Paste network path to Root Path',
-            'Perform update policy: gpupdate/force'
-          ]
-        }
-      }
-    },
-    {
-      id: 'module4',
-      title: 'Server Security and Testing',
-      description: 'Implementing security measures and testing procedures',
-      duration: '8 hours',
-      topics: ['Security Policies', 'Firewall Configuration', 'Testing Procedures', 'Documentation'],
-      practiceContent: {
-        title: 'Domain Joining and Remote Desktop',
-        type: 'B',
-        objectives: 'Join computers to domain, configure network settings, and test remote desktop connections',
-        materials: ['Desktop computer', 'Laptop computer', 'Domain credentials', 'Network configuration'],
-        procedures: [
-          'Configure IP settings on desktop and laptop',
-          'Join computers to domain',
-          'Login with administrator account',
-          'Switch to user accounts',
-          'Test remote desktop connections'
-        ],
-        clientConfiguration: {
-          title: 'Client Computer Configuration',
-          steps: [
-            'Perform ncpa.cpl command',
-            'Right click ethernet adapter, IPv4 properties',
-            'Select Obtain an IP address automatically',
-            'Enter server IP on preferred DNS Server: 10.0.0.5',
-            'Check Status, ensure DHCP Enable shows YES'
-          ]
-        },
-        domainJoining: {
-          title: 'Joining Domain',
-          steps: [
-            'Search and type View Advanced System Settings',
-            'Select Computer Name tab, select Change',
-            'Select Domain and type: "bgfccss.com"',
-            'Enter administrator credentials:',
-            'Username: Administrator, Password: BGFCCSS@2024',
-            'Restart computer after joining'
-          ]
-        },
-        userLogin: {
-          title: 'User Account Login',
-          steps: [
-            'Login using Administrator Account:',
-            'Username: Administrator, Password: BGFCCSS@2024',
-            'Switch user account via CTRL+ALT+DELETE',
-            'Select switch user, enter user credentials:',
-            'Desktop: Username: win10, Password: Admin@123456789',
-            'Laptop: Username: win11, Password: Admin@123456789',
-            'Perform gpupdate/force command',
-            'Test folder redirection configuration'
-          ]
-        },
-        remoteDesktop: {
-          title: 'Remote Desktop Connection',
-          description: 'Perform Remote Desktop connection among SERVER, DESKTOP AND LAPTOP vice versa',
-          steps: [
-            'Open Remote Desktop Connection',
-            'Enter target computer IP address',
-            'Enter administrator credentials',
-            'Test connectivity between all devices',
-            'Verify folder redirection and group policy settings'
-          ]
-        }
-      }
-    }
+    'Configure local server settings and network configuration',
+    'Install and configure server roles including Active Directory',
+    'Set up folder redirection and user account management',
+    'Enable and manage remote desktop connections',
+    'Implement proper server security measures',
+    'Test and validate server functionality'
   ]
 
   const serverTypes = [
     {
-      name: 'File Server',
-      description: 'Manages and shares files across the network',
-      icon: FileText,
-      features: ['File sharing', 'Access control', 'Backup management']
-    },
-    {
-      name: 'Web Server',
-      description: 'Hosts websites and web applications',
-      icon: Server,
-      features: ['HTTP/HTTPS', 'SSL certificates', 'Content management']
-    },
-    {
-      name: 'Database Server',
-      description: 'Manages and stores database information',
-      icon: Database,
-      features: ['Data storage', 'Query processing', 'Data backup']
-    },
-    {
       name: 'Domain Controller',
       description: 'Manages network authentication and directory services',
       icon: Shield,
-      features: ['User authentication', 'Group policies', 'Active Directory']
+      features: ['User authentication', 'Group policies', 'Active Directory', 'DNS services']
+    },
+    {
+      name: 'File Server',
+      description: 'Manages and shares files across the network',
+      icon: FileText,
+      features: ['File sharing', 'Access control', 'Backup management', 'Folder redirection']
+    },
+    {
+      name: 'DHCP Server',
+      description: 'Automatically assigns IP addresses to network devices',
+      icon: Database,
+      features: ['IP address allocation', 'Subnet configuration', 'Lease management', 'Scope management']
+    },
+    {
+      name: 'Remote Desktop Server',
+      description: 'Enables remote access to server and client machines',
+      icon: Monitor,
+      features: ['Remote access', 'Session management', 'Security policies', 'Connection control']
     }
   ]
 
   const networkServices = [
     {
-      name: 'DHCP (Dynamic Host Configuration Protocol)',
-      description: 'Automatically assigns IP addresses to network devices',
-      features: ['IP address allocation', 'Subnet configuration', 'Lease management']
+      name: 'Active Directory Domain Services (AD DS)',
+      description: 'Centralized directory service for user and computer management',
+      features: ['User authentication', 'Group policy management', 'Organizational units', 'Domain services']
     },
     {
       name: 'DNS (Domain Name System)',
       description: 'Translates domain names to IP addresses',
-      features: ['Name resolution', 'Zone management', 'Record configuration']
+      features: ['Name resolution', 'Zone management', 'Record configuration', 'Reverse lookup']
     },
     {
-      name: 'File Services',
-      description: 'Provides file sharing and storage solutions',
-      features: ['Shared folders', 'Permissions', 'Quota management']
+      name: 'DHCP (Dynamic Host Configuration Protocol)',
+      description: 'Automatically assigns IP addresses to network devices',
+      features: ['IP address allocation', 'Subnet configuration', 'Lease management', 'Scope configuration']
     },
     {
-      name: 'Print Services',
-      description: 'Manages network printing and print queues',
-      features: ['Printer sharing', 'Queue management', 'Driver installation']
+      name: 'Folder Redirection',
+      description: 'Redirects user folders to network locations',
+      features: ['Desktop redirection', 'Documents redirection', 'Group policy integration', 'Offline access']
     }
   ]
 
   const securityMeasures = [
     {
-      title: 'Access Control',
-      description: 'Implement user authentication and authorization',
-      icon: Shield
-    },
-    {
       title: 'Firewall Configuration',
-      description: 'Set up network security and traffic filtering',
+      description: 'Configure Windows Firewall to allow necessary services',
       icon: Shield
     },
     {
-      title: 'Data Encryption',
-      description: 'Protect sensitive data with encryption',
-      icon: Shield
+      title: 'User Access Control',
+      description: 'Implement proper user authentication and authorization',
+      icon: Users
     },
     {
-      title: 'Backup and Recovery',
-      description: 'Implement regular backup and disaster recovery',
-      icon: Database
+      title: 'Remote Desktop Security',
+      description: 'Secure remote desktop connections with proper credentials',
+      icon: Monitor
+    },
+    {
+      title: 'Group Policy Security',
+      description: 'Apply security policies through Group Policy Objects',
+      icon: Settings
     }
   ]
 
   const assessmentCriteria = [
-    'Proper server installation and configuration',
-    'Correct user access setup and management',
-    'Successful network services configuration',
-    'Implementation of security measures',
-    'Server testing and validation procedures',
-    'Complete documentation and reporting'
+    'Proper local server configuration and network setup',
+    'Successful installation and configuration of server roles',
+    'Correct folder redirection and user account management',
+    'Effective remote desktop management and connectivity',
+    'Implementation of security measures and policies',
+    'Complete testing and validation of server functionality'
   ]
 
   return (
@@ -346,14 +164,14 @@ export default function COC3Page() {
           <div className="flex items-center mb-6">
             <Server className="h-12 w-12 mr-4" />
             <div>
-              <h1 className="text-4xl font-bold">COC 3: Setting up Computer Servers</h1>
-              <p className="text-xl opacity-90">Basic Server Administration</p>
+              <h1 className="text-4xl font-bold">COC 3: Configure Server</h1>
+              <p className="text-xl opacity-90">Server Configuration and Management</p>
             </div>
           </div>
           <p className="text-lg max-w-3xl">
-            This module covers setting up computer servers including user access configuration, 
-            network services setup, and basic server administration. You will learn to configure 
-            servers, manage users, and implement security measures.
+            This module covers configuring computer servers including local server setup, 
+            Active Directory installation, folder redirection, and remote desktop management. 
+            You will learn to configure servers for domain readiness and network services.
           </p>
         </div>
       </section>
@@ -367,7 +185,7 @@ export default function COC3Page() {
                 <Clock className="h-8 w-8 text-purple-600" />
               </div>
               <h3 className="text-xl font-semibold text-gray-900 mb-2">Duration</h3>
-              <p className="text-gray-600">38 Hours</p>
+              <p className="text-gray-600">30 Hours</p>
             </div>
             <div className="text-center">
               <div className="bg-blue-100 w-16 h-16 rounded-full flex items-center justify-center mx-auto mb-4">
@@ -500,287 +318,7 @@ export default function COC3Page() {
       </section>
 
       {/* Modules */}
-      <section className="py-16 bg-gray-50">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="text-center mb-12">
-            <h2 className="text-3xl font-bold text-gray-900 mb-4">Course Modules</h2>
-            <p className="text-lg text-gray-600 max-w-3xl mx-auto">
-              Structured learning modules covering all aspects of server setup and administration.
-            </p>
-          </div>
-          
-          <div className="space-y-6">
-            {modules.map((module, index) => (
-              <div key={module.id} className="bg-white rounded-lg p-6 shadow-md">
-                <div className="flex items-start justify-between">
-                  <div className="flex-1">
-                    <div className="flex items-center mb-3">
-                      <span className="bg-purple-600 text-white w-8 h-8 rounded-full flex items-center justify-center text-sm font-bold mr-4">
-                        {index + 1}
-                      </span>
-                      <h3 className="text-xl font-semibold text-gray-900">{module.title}</h3>
-                    </div>
-                    <p className="text-gray-600 mb-4">{module.description}</p>
-                    <div className="flex items-center text-sm text-gray-500 mb-4">
-                      <Clock className="h-4 w-4 mr-1" />
-                      {module.duration}
-                    </div>
-                    <div className="grid md:grid-cols-2 gap-2">
-                      {module.topics.map((topic, topicIndex) => (
-                        <div key={topicIndex} className="flex items-center text-sm text-gray-600">
-                          <ChevronRight className="h-3 w-3 mr-2 text-purple-500" />
-                          {topic}
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                  <div className="ml-6 flex space-x-2">
-                    <button className="bg-purple-600 text-white px-4 py-2 rounded-lg hover:bg-purple-700 transition-colors flex items-center">
-                      <Video className="h-4 w-4 mr-2" />
-                      Watch
-                    </button>
-                    <button 
-                      onClick={() => {
-                        const modal = document.getElementById(`practice-modal-${module.id}`);
-                        if (modal) modal.classList.remove('hidden');
-                      }}
-                      className="bg-blue-600 text-white px-4 py-2 rounded-lg hover:bg-blue-700 transition-colors flex items-center"
-                    >
-                      <FileText className="h-4 w-4 mr-2" />
-                      Practice
-                    </button>
-                  </div>
-                </div>
-
-                {/* Practice Modal */}
-                <div id={`practice-modal-${module.id}`} className="hidden fixed inset-0 bg-black bg-opacity-50 z-50 flex items-center justify-center p-4">
-                  <div className="bg-white rounded-lg max-w-4xl w-full max-h-[90vh] overflow-y-auto">
-                    <div className="p-6">
-                      <div className="flex justify-between items-center mb-6">
-                        <h3 className="text-2xl font-bold text-gray-900">{module.practiceContent?.title}</h3>
-                        <button 
-                          onClick={() => {
-                            const modal = document.getElementById(`practice-modal-${module.id}`);
-                            if (modal) modal.classList.add('hidden');
-                          }}
-                          className="text-gray-500 hover:text-gray-700"
-                        >
-                          <svg className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                          </svg>
-                        </button>
-                      </div>
-
-                      {module.practiceContent && (
-                        <div className="space-y-6">
-                          {/* Objectives */}
-                          <div className="bg-purple-50 rounded-lg p-4">
-                            <h4 className="font-semibold text-purple-900 mb-2">Objectives:</h4>
-                            <p className="text-purple-800">{module.practiceContent.objectives}</p>
-                          </div>
-
-                          {/* Materials */}
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Materials Needed:</h4>
-                            <ul className="list-disc list-inside space-y-1 text-gray-700">
-                              {module.practiceContent.materials.map((material, idx) => (
-                                <li key={idx}>{material}</li>
-                              ))}
-                            </ul>
-                          </div>
-
-                          {/* Procedures */}
-                          <div>
-                            <h4 className="font-semibold text-gray-900 mb-3">Step-by-Step Procedures:</h4>
-                            <ol className="list-decimal list-inside space-y-2 text-gray-700">
-                              {module.practiceContent.procedures.map((procedure, idx) => (
-                                <li key={idx}>{procedure}</li>
-                              ))}
-                            </ol>
-                          </div>
-
-                          {/* Firewall Settings for Module 1 */}
-                          {module.id === 'module1' && module.practiceContent.firewallSettings && (
-                            <div>
-                              <h4 className="font-semibold text-gray-900 mb-3">Firewall Configuration:</h4>
-                              <div className="bg-blue-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-blue-900 mb-2">{module.practiceContent.firewallSettings.title}</h5>
-                                <p className="text-blue-800 mb-3">{module.practiceContent.firewallSettings.description}</p>
-                                <ul className="space-y-1">
-                                  {module.practiceContent.firewallSettings.services.map((service, idx) => (
-                                    <li key={idx} className="flex items-center text-blue-700">
-                                      <CheckCircle className="h-4 w-4 mr-2" />
-                                      {service}
-                                    </li>
-                                  ))}
-                                </ul>
-                              </div>
-                            </div>
-                          )}
-
-                          {/* Network Configuration for Module 1 */}
-                          {module.id === 'module1' && module.practiceContent.networkConfiguration && (
-                            <div>
-                              <h4 className="font-semibold text-gray-900 mb-3">Network Configuration:</h4>
-                              <div className="bg-green-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-green-900 mb-2">{module.practiceContent.networkConfiguration.title}</h5>
-                                <div className="grid md:grid-cols-2 gap-4">
-                                  {Object.entries(module.practiceContent.networkConfiguration.settings).map(([key, value]) => (
-                                    <div key={key} className="flex justify-between">
-                                      <span className="font-medium text-green-800">{key}:</span>
-                                      <span className="text-green-700">{value}</span>
-                                    </div>
-                                  ))}
-                                </div>
-                              </div>
-                            </div>
-                          )}
-
-                          {/* Active Directory Setup for Module 2 */}
-                          {module.id === 'module2' && module.practiceContent.activeDirectorySetup && (
-                            <div>
-                              <h4 className="font-semibold text-gray-900 mb-3">Active Directory Setup:</h4>
-                              <div className="bg-yellow-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-yellow-900 mb-2">{module.practiceContent.activeDirectorySetup.title}</h5>
-                                <ol className="list-decimal list-inside space-y-2 text-yellow-800">
-                                  {module.practiceContent.activeDirectorySetup.steps.map((step, idx) => (
-                                    <li key={idx}>{step}</li>
-                                  ))}
-                                </ol>
-                              </div>
-                            </div>
-                          )}
-
-                          {/* DHCP Configuration for Module 2 */}
-                          {module.id === 'module2' && module.practiceContent.dhcpConfiguration && (
-                            <div>
-                              <h4 className="font-semibold text-gray-900 mb-3">DHCP Configuration:</h4>
-                              <div className="bg-purple-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-purple-900 mb-2">{module.practiceContent.dhcpConfiguration.title}</h5>
-                                <ol className="list-decimal list-inside space-y-2 text-purple-800">
-                                  {module.practiceContent.dhcpConfiguration.steps.map((step, idx) => (
-                                    <li key={idx}>{step}</li>
-                                  ))}
-                                </ol>
-                              </div>
-                            </div>
-                          )}
-
-                          {/* User Creation for Module 2 */}
-                          {module.id === 'module2' && module.practiceContent.userCreation && (
-                            <div>
-                              <h4 className="font-semibold text-gray-900 mb-3">User Creation:</h4>
-                              <div className="bg-blue-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-blue-900 mb-2">{module.practiceContent.userCreation.title}</h5>
-                                <ol className="list-decimal list-inside space-y-2 text-blue-800">
-                                  {module.practiceContent.userCreation.steps.map((step, idx) => (
-                                    <li key={idx}>{step}</li>
-                                  ))}
-                                </ol>
-                              </div>
-                            </div>
-                          )}
-
-                          {/* Folder Sharing for Module 3 */}
-                          {module.id === 'module3' && module.practiceContent.folderSharing && (
-                            <div>
-                              <h4 className="font-semibold text-gray-900 mb-3">Folder Sharing:</h4>
-                              <div className="bg-green-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-green-900 mb-2">{module.practiceContent.folderSharing.title}</h5>
-                                <ol className="list-decimal list-inside space-y-2 text-green-800">
-                                  {module.practiceContent.folderSharing.steps.map((step, idx) => (
-                                    <li key={idx}>{step}</li>
-                                  ))}
-                                </ol>
-                              </div>
-                            </div>
-                          )}
-
-                          {/* Group Policy for Module 3 */}
-                          {module.id === 'module3' && module.practiceContent.groupPolicy && (
-                            <div>
-                              <h4 className="font-semibold text-gray-900 mb-3">Group Policy Configuration:</h4>
-                              <div className="bg-purple-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-purple-900 mb-2">{module.practiceContent.groupPolicy.title}</h5>
-                                <ol className="list-decimal list-inside space-y-2 text-purple-800">
-                                  {module.practiceContent.groupPolicy.steps.map((step, idx) => (
-                                    <li key={idx}>{step}</li>
-                                  ))}
-                                </ol>
-                              </div>
-                            </div>
-                          )}
-
-                          {/* Client Configuration for Module 4 */}
-                          {module.id === 'module4' && module.practiceContent.clientConfiguration && (
-                            <div>
-                              <h4 className="font-semibold text-gray-900 mb-3">Client Configuration:</h4>
-                              <div className="bg-blue-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-blue-900 mb-2">{module.practiceContent.clientConfiguration.title}</h5>
-                                <ol className="list-decimal list-inside space-y-2 text-blue-800">
-                                  {module.practiceContent.clientConfiguration.steps.map((step, idx) => (
-                                    <li key={idx}>{step}</li>
-                                  ))}
-                                </ol>
-                              </div>
-                            </div>
-                          )}
-
-                          {/* Domain Joining for Module 4 */}
-                          {module.id === 'module4' && module.practiceContent.domainJoining && (
-                            <div>
-                              <h4 className="font-semibold text-gray-900 mb-3">Domain Joining:</h4>
-                              <div className="bg-green-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-green-900 mb-2">{module.practiceContent.domainJoining.title}</h5>
-                                <ol className="list-decimal list-inside space-y-2 text-green-800">
-                                  {module.practiceContent.domainJoining.steps.map((step, idx) => (
-                                    <li key={idx}>{step}</li>
-                                  ))}
-                                </ol>
-                              </div>
-                            </div>
-                          )}
-
-                          {/* User Login for Module 4 */}
-                          {module.id === 'module4' && module.practiceContent.userLogin && (
-                            <div>
-                              <h4 className="font-semibold text-gray-900 mb-3">User Login:</h4>
-                              <div className="bg-yellow-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-yellow-900 mb-2">{module.practiceContent.userLogin.title}</h5>
-                                <ol className="list-decimal list-inside space-y-2 text-yellow-800">
-                                  {module.practiceContent.userLogin.steps.map((step, idx) => (
-                                    <li key={idx}>{step}</li>
-                                  ))}
-                                </ol>
-                              </div>
-                            </div>
-                          )}
-
-                          {/* Remote Desktop for Module 4 */}
-                          {module.id === 'module4' && module.practiceContent.remoteDesktop && (
-                            <div>
-                              <h4 className="font-semibold text-gray-900 mb-3">Remote Desktop Connection:</h4>
-                              <div className="bg-purple-50 rounded-lg p-4">
-                                <h5 className="font-semibold text-purple-900 mb-2">{module.practiceContent.remoteDesktop.title}</h5>
-                                <p className="text-purple-800 mb-3">{module.practiceContent.remoteDesktop.description}</p>
-                                <ol className="list-decimal list-inside space-y-2 text-purple-800">
-                                  {module.practiceContent.remoteDesktop.steps.map((step, idx) => (
-                                    <li key={idx}>{step}</li>
-                                  ))}
-                                </ol>
-                              </div>
-                            </div>
-                          )}
-                        </div>
-                      )}
-                    </div>
-                  </div>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      </section>
+      <COC3Modules />
 
       {/* Assessment Criteria */}
       <section className="py-16 bg-white">
@@ -829,4 +367,4 @@ export default function COC3Page() {
       <Footer />
     </div>
   )
-} 
+}
