+++ src/app/contact/page.tsx
@@ -1,6 +1,7 @@
 "use client"
 
 import Link from 'next/link'
+import { Metadata } from 'next'
 import { 
   ArrowLeft, 
   Mail, 
@@ -13,6 +14,34 @@ import {
   MessageSquare
 } from 'lucide-react'
 import Footer from '@/components/Footer'
+
+export const metadata: Metadata = {
+  title: 'Contact Us - CSS NC II Training Companion App',
+  description: 'Get in touch with the CSS NC II Training Companion App team. Contact us for support, questions, or feedback about our Computer Systems Servicing NC II training program.',
+  keywords: 'Contact, CSS NC II, Training Support, Computer Systems Servicing, TESDA, Training Questions, Feedback',
+  openGraph: {
+    title: 'Contact Us - CSS NC II Training Companion App',
+    description: 'Get in touch with the CSS NC II Training Companion App team. Contact us for support, questions, or feedback.',
+    url: 'https://css-nc2-companion.vercel.app/contact',
+    siteName: 'CSS NC II Training Companion App',
+    images: [
+      {
+        url: '/img/facebook_meta_bg.jpg',
+        width: 1200,
+        height: 630,
+        alt: 'Contact CSS NC II Training Companion App',
+      },
+    ],
+    locale: 'en_US',
+    type: 'website',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'Contact Us - CSS NC II Training Companion App',
+    description: 'Get in touch with the CSS NC II Training Companion App team. Contact us for support, questions, or feedback.',
+    images: ['/img/facebook_meta_bg.jpg'],
+  },
+};
 
 export default function ContactPage() {
   const handleSubmit = (e: React.FormEvent) => {
